Removing path and file manipulation from the core function
A library should not display graphic interface. This function should
return more results or write more data to disk.

- make the display False by default
- write graph to disk if afficher_hist=False

Seems that fourcc codec depends on the platform and the video type. At
the moment, it is hard coded in peak() function :
https://docs.opencv.org/4.x/dd/d43/tutorial_py_video_display.html

The videos used in examples did not exist in samples folder.

Used a video ot the sample folder but the parameters are hard coded.