Kept only 2 variables PATH_IN and PATH_OUT
There was no clear logic between PATH_IN and VIDEO_PATH_IN for example.
This can be changed again with a better logic.

Removed the trailing "/" if not necessary.

To be robust to "../somewhere-else", we may have to call resolve but it
has to be tested to be sure there are no drawbacks.